+++ app/models/recommendation_track.rb
@@ -1,0 +1,5 @@
+class RecommendationTrack < ApplicationRecord
+  belongs_to :recommendation_collection
+
+  validates :item_id, :position, :title, presence: true
+end