+++ app/models/server_connection.rb
@@ -3,6 +3,7 @@ class ServerConnection < ApplicationRecord
 
   belongs_to :user
   belongs_to :media_server
+  has_many :listening_events, dependent: :destroy
 
   validates :username, presence: true
   validates :access_token, presence: true, unless: :password?