+++ app/models/user.rb
@@ -2,6 +2,9 @@ class User < ApplicationRecord
   has_secure_password
   has_many :sessions, dependent: :destroy
   has_many :server_connections, dependent: :destroy
+  has_many :recommendation_collections, dependent: :destroy
+  has_many :recommendation_runs, dependent: :destroy
+  has_many :listening_events, dependent: :destroy
 
   normalizes :email_address, with: ->(e) { e.strip.downcase }