+++ config/routes.rb
@@ -9,6 +9,9 @@ Rails.application.routes.draw do
   root "home#index"
   get "offline_shell", to: "offline_shells#show", as: :offline_shell
   resources :offline_downloads, only: :index
+  resources :recommendation_collections, only: %i[index show] do
+    post :events, on: :member
+  end
   get "home/content", to: "home#content", as: :home_content
   get "library/artists", to: "library#artists", as: :library_artists
   get "library/albums", to: "library#albums", as: :library_albums